Fixed: 2 bad cast warning
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@214 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/src/3d/tessellation.cpp
@@ -711,8 +711,8 @@ void		CTessFace::computeTileMaterial()
 	{
 		sint	i;
 		TileMaterial= Patch->getLandscape()->newTileMaterial();
-		TileMaterial->TileS= ts;
-		TileMaterial->TileT= tt;
+		TileMaterial->TileS= uint8(ts);
+		TileMaterial->TileT= uint8(tt);
 
 		// Add this new material to the render list.
 		Patch->appendTileMaterialToRenderList(TileMaterial);
